more refactor 2
